backend/requirements.txt
Django
djangorestframework
django-cors-headers
python-dotenv
psycopg[binary]
Pillow
djangorestframework-simplejwt
environs
whitenoise
django-filter
python-telegram-bot
cryptography